backend/requirements.txt
openai
python-dotenv
sendgrid
tavily-python
markdown
pydantic
openai-agents
fastapi
uvicorn[standard]
